+++ src/plugins/intel_cpu/src/emitters/plugin/aarch64/jit_load_store_emitters.hpp
@@ -38,9 +38,6 @@ private:
     template <dnnl::impl::cpu::aarch64::cpu_isa_t isa>
     void load_byte(const std::vector<size_t> &in_idxs, const std::vector<size_t> &out_idxs) const;
     size_t get_aux_gprs_count() const override;
-    size_t get_aux_vecs_count() const override;
-
-    std::unique_ptr<jit_convert_truncation_emitter> convert_truncation_emitter = nullptr;
 
     std::string name_;
     int load_num_;  // the element number to load
@@ -69,9 +66,6 @@ private:
     template <dnnl::impl::cpu::aarch64::cpu_isa_t isa>
     void store_byte(const std::vector<size_t> &in_idxs, const std::vector<size_t> &out_idxs) const;
     size_t get_aux_gprs_count() const override;
-    size_t get_aux_vecs_count() const override;
-
-    std::unique_ptr<jit_convert_emitter> convert_emitter = nullptr;
 
     std::string name_;
     int store_num_;  // the element number to store
